% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartNP-package.R
\docType{package}
\name{bartNP-package}
\alias{bartNP}
\alias{bartNP-package}
\title{bartNP: Execute BART on TF perturbation data}
\description{
Extend SummarizedExperiment with slots and functions to ease TF effect modelling with BART
}
\author{
\strong{Maintainer}: Chase Mateusiak \email{chasem@wustl.edu} (\href{https://orcid.org/0000-0002-2890-4242}{ORCID})

}
\keyword{internal}
