% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{exprMat<-}
\alias{exprMat<-}
\alias{exprMat}
\alias{exprMat<-,NetProphetDataSet-method}
\alias{exprMat,NetProphetDataSet-method}
\title{Set expression matrix}
\usage{
exprMat(x, ...) <- value

exprMat(x, ...)

\S4method{exprMat}{NetProphetDataSet}(x) <- value

\S4method{exprMat}{NetProphetDataSet}(x)
}
\arguments{
\item{x}{a NetProphetDataSet object}

\item{value}{an expression matrix with columns equal to colData, and colnames
equal to and in the same order as the colnames column of colData}
}
\description{
Set or replace the expression matrix in the \link{NetProphetDataSet}

Extract the expression matrix from the NetProphetDataSet
}
