% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{nonRegGenes}
\alias{nonRegGenes}
\alias{nonRegGenes,NetProphetDataSet-method}
\title{get list of genes from the rownames of the \link{NetProphetDataSet} which are not
regulators, which are those genes in the columns of the regMatrix slot}
\usage{
nonRegGenes(x, ...)

\S4method{nonRegGenes}{NetProphetDataSet}(x)
}
\arguments{
\item{x}{a NetProphetDataSet object}
}
\description{
get list of genes from the rownames of the \link{NetProphetDataSet} which are not
regulators, which are those genes in the columns of the regMatrix slot
}
