% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{regMatrix<-}
\alias{regMatrix<-}
\alias{regMatrix}
\alias{regMatrix<-,NetProphetDataSet-method}
\alias{regMatrix,NetProphetDataSet-method}
\title{gene by regulators matrix}
\usage{
regMatrix(x, ...) <- value

regMatrix(x, ...)

\S4method{regMatrix}{NetProphetDataSet}(x) <- value

\S4method{regMatrix}{NetProphetDataSet}(x)
}
\arguments{
\item{x}{a NetProphetDataSet object}

\item{value}{a gene by regulation matrix, where the number of genes (rows) is
equal to the rows in the \link{NetProphetDataSet} expression matrix}
}
\description{
A matrix with dimension genes, which is the same as the expression matrix of
the \link{NetProphetDataSet}, and columns corresponding to regulators. Values are
boolean values where \code{TRUE} allows regulator j to affect the transcription of
gene i, and \code{FALSE} disallows regulation.
}
