% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{regPredictors}
\alias{regPredictors}
\alias{regPredictors,NetProphetDataSet-method}
\title{extract a subset of the expression matrix filtered by the regulation matrix
for a given gene}
\usage{
regPredictors(x, ...)

\S4method{regPredictors}{NetProphetDataSet}(x, gene, transpose = TRUE)
}
\arguments{
\item{x}{a NetProphetDataSet object}

\item{gene}{a gene which is in the rownames(x) and possibly in the colnames(regMatrix(x))}

\item{transpose}{logical, default TRUE to return a sample x predictor matrix.
NOTE that this is of particular interest when calling C++ (and maybe C?)
functions as they store matricies differently than R and Fortran (row vs column wise).
Best to check on these details if you think it might matter.}
}
\description{
This function returns a matrix in dimensions samples, where samples are the
columns of the \link{NetProphetDataSet}, x predictors, where predictors are a
subset of the regulators. Any regulators which do not regulate a gene
according to the regulation matrix are filtered out
}
