% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[,NetProphetDataSet,ANY,ANY,ANY-method}
\alias{[,NetProphetDataSet,ANY,ANY,ANY-method}
\title{subsetting the NetProphetDataSet}
\usage{
\S4method{[}{NetProphetDataSet,ANY,ANY,ANY}(x, i, j, drop = TRUE)
}
\arguments{
\item{x}{a NetProphetDataSet object}

\item{i}{column index or incicies (a numeric or boolean vector)}

\item{j}{column filter -- an index, name, boolean or vector of them}

\item{drop}{default TRUE. This may not do anything -- need to check if
it is part of the Extending SummarizedExperiment documentation, or a
feature I added, then removed and forgot to remove the argument.}
}
\value{
a NetProphetDataSet. If the subsetting is on the genes, the regulation
matrix will also be gene filtered in the same way as the counts. Column
filtering does not currently affect the regulation matrix
}
\description{
subsetting the NetProphetDataSet
}
