% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NetProphetDataSet-class}
\alias{NetProphetDataSet-class}
\alias{.NetProphetDataSet}
\title{A container for expression data meant to be used to infer regulator relationships}
\description{
A container for expression data meant to be used to infer regulator relationships
}
\seealso{
\itemize{
\item \href{https://www.bioconductor.org/packages/devel/bioc/vignettes/SummarizedExperiment/inst/doc/Extensions.html}{extending SummarizedExperiment}.
Note that A \link{SummarizedExperiment} object becomes a \link{RangedSummarizedExperiment}
object when rowRanges are added.
}
}
